requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
web3==6.11.3
requests==2.31.0
Werkzeug==3.0.1
